using 8-byte variables

+++ include/Graph.hpp
@@ -49,5 +49,5 @@ public:
 
     uint32_t find_first_attainable_year();
 
-    uint32_t find_lowest_possible_cost();
+    uint64_t find_lowest_possible_cost();
 };
