+++ include/Village.hpp
@@ -11,9 +11,9 @@
 class Village {
 private:
     uint32_t m_index_village;
-    uint32_t m_age_conn;
-    uint32_t m_crossing_time;
-    uint32_t m_construction_cost;
+    uint64_t m_age_conn;
+    uint64_t m_crossing_time;
+    uint64_t m_construction_cost;
     bool m_visited;
     int32_t m_parent_village;
     std::list<std::shared_ptr<Connection>> m_connected_villages;
@@ -27,11 +27,11 @@ public:
 
     uint32_t get_index_village() const;
 
-    uint32_t get_age_conn() const;
+    uint64_t get_age_conn() const;
 
-    uint32_t get_crossing_time() const;
+    uint64_t get_crossing_time() const;
 
-    uint32_t get_construction_cost() const;
+    uint64_t get_construction_cost() const;
 
     bool has_visited() const;
 
@@ -41,11 +41,11 @@ public:
 
     void set_index_village(uint32_t index_village);
 
-    void set_age_conn(uint32_t age_conn);
+    void set_age_conn(uint64_t age_conn);
 
-    void set_crossing_time(uint32_t crossing_time);
+    void set_crossing_time(uint64_t crossing_time);
 
-    void set_construction_cost(uint32_t construction_cost);
+    void set_construction_cost(uint64_t construction_cost);
 
     void set_visited(bool visit);
